chore: refactor to --enrich flag
